requirements.txt
Flask[async] # Changed this line
PyPDF2
python-dotenv
chromadb
google-generativeai
Flask-CORS